KTOR-4976 migrate to new testing API
